feat: Переход на очеред команд с сервера вместо callback. Удалены CAS команды. Глобальный рефакторинг API для работы с клиентом

+++ client/Unity/Assets/Shared/PlayerHelper.cs
@@ -2,6 +2,6 @@ namespace Shared
 {
     public static class PlayerHelper
     {
-        public const ulong PlayerGroup = 8;
+        public const ushort PlayerGroup = 8;
     }
 }
